refactored UI parts of new functionality
