+++ src/components/dashboard/resourcesPage/Resources.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import ResourceItem from "./ResourceItem";
+import ResourceItem from "./resourceItem/ResourceItem";
 
 interface Resources {
   resource_id: string;
@@ -9,7 +9,7 @@ interface Resources {
 
 export default function Resources({ data }: { data: any }) {
   return (
-    <div className="rounded flex flex-wrap gap-5 shadow-lg m-4 justify-center lg:justify-normal">
+    <div className="w-full grid grid-cols-1 sm:grid-cols-2 xl:grid-cols-4 gap-5">
       {data.map((item: any) => {
         return <ResourceItem item={item} key={item.resource_id} />;
       })}
