+++ src/components/dashboard/resourcesPage/resourceItem/actions/DecreaseCount.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { actDecreaseCount } from "../../../lib/actions";
+import { actDecreaseCount } from "../../../../../lib/actions";
 
 export default function DecreaseCount({ id }: { id: string }) {
   return (
@@ -8,7 +8,7 @@ export default function DecreaseCount({ id }: { id: string }) {
       className="border border-gray-300 p-2 rounded-lg"
       onClick={() => actDecreaseCount(id)}
     >
-      -1
+      - 1
     </button>
   );
 }
