+++ src/components/dashboard/resourcesPage/resourceItem/actions/ItemCount.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+export default async function ItemCount({ count }: { count: string }) {
+  if (count) {
+    return <p>{count}</p>;
+  } else {
+    return null;
+  }
+}
