--- src/components/ui/bookmarkAction/Title.tsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export default function Title({ title }: { title: any }) {
-  return <>{title ? "remove from" : "add to"} bookmark</>;
-}
